Victron Smart BatteryProtect file renaming
