+++ src/devices.h
@@ -115,8 +115,8 @@
 #include "devices/SensorP_HT_json.h"
 #include "devices/ServiceData_json.h"
 #include "devices/UT363BT_json.h"
-#include "devices/VICTRON_DCDC_ENCR_json.h"
-#include "devices/VICTRON_DCDC_json.h"
+#include "devices/VICTRON_SBP_ENCR_json.h"
+#include "devices/VICTRON_SBP_json.h"
 #include "devices/VICTRON_ORIONXS_ENCR_json.h"
 #include "devices/VICTRON_ORIONXS_json.h"
 #include "devices/SE_RHT_json.h"
